Drop dependency on p2
Prior to this change, the test-smoke script required the p2 binary to
augment the Dockerfile under docker-postgres-replication.  I do not have
that binary on my system, nor was I able to find an installer for it.
The closest tool I found was p2cli, but even that did not work as
expected.

This change removes the Dockerfile.p2 file in favour of using Docker's
builtin build args mechanism to augment the docker build.

Signed-off-by: Corin Lawson <corin@responsight.com>